requirements.txt
azure-common==1.1.23
azure-core==1.1.1
azure-functions==1.0.7
azure-nspkg==3.0.2
azure-storage-blob==12.1.0
azure-storage-nspkg==3.1.0
